Revert "Scripts/HallsOfReflection: fix some warning / whitespace / style"
This reverts commit 3cc2619f374c2bc7d289a5010728f4e7b161b329.

+++ src/server/scripts/Northrend/FrozenHalls/HallsOfReflection/boss_marwyn.cpp
@@ -82,7 +82,7 @@ public:
 
        uint8 SummonCount;
 
-       uint32 pSummon;
+       uint64 pSummon;
 
        void Reset()
        {
